HookFormInputStatus to InputStatus
