--- src/components/shared/Fields/InputStatus/HookForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InputStatus';
